Remove deprecated quarter interval and timezone tests; consolidate into a single test file for improved organization and maintainability.
